+++ block/src/switch/woocommerce-tm-extra-product-options.js
@@ -1,0 +1,17 @@
+import WoocommerceTmExtraProductOptions from './compatibility/WoocommerceTmExtraProductOptions';
+
+window.addEventListener( 'DOMContentLoaded', () => {
+	const viewConfig = window.wtsViewObject || {
+		originalTaxDisplay: 'incl',
+	};
+
+	const wtsCompatibilityObject = window.wtsCompatibilityObject || {
+		baseTaxRate: 0,
+	};
+
+	const tmtepo = new WoocommerceTmExtraProductOptions(
+		viewConfig.originalTaxDisplay,
+		wtsCompatibilityObject.baseTaxRate
+	);
+	tmtepo.init();
+} );