+++ wdevs-tax-switch.php
@@ -16,7 +16,7 @@
  * Plugin Name:          Tax Switch for WooCommerce
  * Plugin URI:           https://wijnberg.dev
  * Description:          Let customers toggle between inclusive and exclusive VAT pricing in your WooCommerce store.
- * Version:              1.5.11
+ * Version:              1.5.12
  * Author:               Wijnberg Developments
  * Author URI:           https://wijnberg.dev/
  * License:              GPL-2.0+
@@ -26,7 +26,7 @@
  * Requires PHP:         7.2
  * Requires at least:    5.0
  * WC requires at least: 7.0.0
- * WC tested up to:      10.0.2
+ * WC tested up to:      10.1.0
  * Requires Plugins:     woocommerce
  */
 
@@ -40,7 +40,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'WDEVS_TAX_SWITCH_VERSION', '1.5.11' );
+define( 'WDEVS_TAX_SWITCH_VERSION', '1.5.12' );
 
 /**
  * The code that runs during plugin activation.